refactor: 테스트 코드 작성을 위한 서비스 추상화
